NXP-25369: Try to fix hot reload test errors
